move GTSFMProcess out of registry file
