+++ gtsfm/two_view_estimator.py
@@ -35,7 +35,7 @@ from gtsfm.frontend.inlier_support_processor import InlierSupportProcessor
 from gtsfm.frontend.matcher.matcher_base import MatcherBase
 from gtsfm.frontend.verifier.verifier_base import VerifierBase
 from gtsfm.evaluation.metrics import GtsfmMetric, GtsfmMetricsGroup
-from gtsfm.ui.registry import GTSFMProcess, UiMetadata
+from gtsfm.ui.gtsfm_process import GTSFMProcess, UiMetadata
 
 
 logger = logger_utils.get_logger()
@@ -56,7 +56,7 @@ class TwoViewEstimator(GTSFMProcess):
     """Wrapper for running two-view relative pose estimation on image pairs in the dataset."""
 
     def get_ui_metadata() -> UiMetadata:
-        """Returns data needed to display this process in the process graph. See gtsfm/ui/registry.py for more info."""
+        """Returns data needed to display this process in the process graph."""
 
         return UiMetadata(
             "Two-View Estimator",
